src/app/requirements.txt
streamlit==1.40.2
pydantic==2.7.1
python-dotenv==1.0.0
azure-ai-projects==1.0.0b4
azure-identity==1.17.1
uvicorn==0.29.0
gunicorn
pandas==2.2.3
numpy==1.24.4
pypdf==4.1.0
azure-core==1.30.0
azure-search-documents==11.4.0
openai==1.55.3
pyodbc==4.0.39